refactor(migrate): the verbs that print a yes/no answer live together
Adding org-exists pushed main.go past the 500-line file ceiling, which is the
named trigger for a split rather than a reason to shrink the addition.

The cut is by concern, not by size: dbExists and orgExists both answer one
question by PRINTING the answer instead of setting an exit code, so a shell
caller can branch on it without conflating "the answer is no" with "the
command failed". That output contract is what they share and what their
callers depend on -- scripts/lib-testdb.sh string-compares one, the deploy
entrypoint string-compares the other.
